move dlk from vendor
