+++ build.sh
@@ -6,7 +6,7 @@ docker build -t ${PREFIX}vizier-core -f manager/Dockerfile .
 docker build -t ${PREFIX}suggestion-random -f suggestion/random/Dockerfile .
 docker build -t ${PREFIX}suggestion-grid -f suggestion/grid/Dockerfile .
 docker build -t ${PREFIX}suggestion-hyperband -f suggestion/hyperband/Dockerfile .
-docker build -t ${PREFIX}dlk-manager -f vendor/github.com/osrg/dlk/build/Dockerfile vendor/github.com/osrg/dlk
+docker build -t ${PREFIX}dlk-manager -f dlk/Dockerfile .
 docker build -t ${PREFIX}katib-frontend -f manager/modeldb/Dockerfile .
 docker build -t ${PREFIX}katib-cli -f cli/Dockerfile .
 mkdir -p bin
