+++ dlk/dlkctl/bash-cmp.go
