+++ dlk/dlkctl/del.go
