+++ dlk/dlkmanager/bash-cmp.go
