+++ dlk/dlkmanager/learningTask.go
@@ -7,8 +7,8 @@ import (
 	"sync"
 	"time"
 
-	"github.com/osrg/dlk/dlkmanager/api"
-	"github.com/osrg/dlk/dlkmanager/datastore"
+	"github.com/kubeflow/hp-tuning/dlk/dlkmanager/api"
+	"github.com/kubeflow/hp-tuning/dlk/dlkmanager/datastore"
 
 	lgr "github.com/sirupsen/logrus"
 	batchv1 "k8s.io/api/batch/v1"
