+++ manager/worker_interface/dlk/dlk.go
@@ -7,9 +7,9 @@ import (
 	"fmt"
 	"github.com/kubeflow/hp-tuning/api"
 	"github.com/kubeflow/hp-tuning/db"
+	dlkapi "github.com/kubeflow/hp-tuning/dlk/dlkmanager/api"
+	"github.com/kubeflow/hp-tuning/dlk/dlkmanager/datastore"
 	"github.com/kubeflow/hp-tuning/manager/modeldb"
-	dlkapi "github.com/osrg/dlk/dlkmanager/api"
-	"github.com/osrg/dlk/dlkmanager/datastore"
 	"io/ioutil"
 	"log"
 	"net/http"
